Extract Item.Preview
