+++ src/commands/item.js
@@ -3,15 +3,8 @@
 const { Command } = require('./command')
 const { SaveCommand } = require('./subject')
 const mod = require('../models')
-const { darwin } = require('../common/os')
 const { ITEM } = require('../constants')
-const { win } = require('../window')
-const { get } = require('../common/util')
-
-const {
-  call,
-  select
-} = require('redux-saga/effects')
+const { call } = require('redux-saga/effects')
 
 
 class Load extends Command {
@@ -35,35 +28,16 @@ class TemplateChange extends SaveCommand {
 TemplateChange.register(ITEM.TEMPLATE.CHANGE)
 
 
-
-
-class Preview extends Command {
-  *exec() {
-    if (!darwin) return
-
-    let { photos } = this.action.payload
-    let paths = yield select(state =>
-      photos.map(id => get(state.photos, [id, 'path'])))
-
-    if (paths.length > 0) {
-      win.preview(paths[0])
-    }
-  }
-}
-
-Preview.register(ITEM.PREVIEW)
-
-
 module.exports = {
   ...require('./item/create'),
   ...require('./item/explode'),
   ...require('./item/export'),
   ...require('./item/import'),
   ...require('./item/merge'),
+  ...require('./item/preview'),
   ...require('./item/print'),
   ...require('./item/tags'),
 
   Load,
-  TemplateChange,
-  Preview
+  TemplateChange
 }
